+++ test/github-issues/2103/entity/Complex.ts
@@ -1,0 +1,15 @@
+import {Entity, Column, PrimaryColumn} from "../../../../src";
+
+@Entity()
+export class Complex {
+
+    @PrimaryColumn()
+    id: number;
+
+    @PrimaryColumn()
+    code: number;
+
+    @Column()
+    x: number;
+
+}