+++ test/github-issues/2103/entity/Simple.ts
@@ -1,0 +1,12 @@
+import {Entity, Column, PrimaryGeneratedColumn} from "../../../../src";
+
+@Entity()
+export class Simple {
+
+    @PrimaryGeneratedColumn()
+    id: number;
+
+    @Column()
+    x: number;
+
+}